nonlogs-full-stack/nonlogs-full-stack/api/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.4.4
python-dotenv==1.0.0
Werkzeug==2.3.6
SQLAlchemy==2.0.19
gunicorn==21.2.0
requests==2.31.0
